requirements.txt

streamlit~=1.45.0
music21~=8.3.0
numpy~=2.0.2
dtw-python
opencv-python
midi2audio~=0.1.1
librosa~=0.9.2
numpy
matplotlib
scipy
dtw-python
fastdtw
